+++ app/utils/urlHelpers.js
@@ -1,4 +1,5 @@
 import { parse } from 'url';
+import path from 'path';
 import pkg from 'appPackage';
 import logger from 'logger';
 import { PROTOCOLS } from 'appConstants';
@@ -9,37 +10,53 @@ export const isInternalPage = ( tab ) =>
     const urlObj = parse( tab.url );
 
     return urlObj.protocol === `${PROTOCOLS.INTERNAL_PAGES}:`;
-}
+};
 
 export const removeTrailingSlash = ( url ) =>
+    url.replace( /\/$/, '' );
+
+
+export const addTrailingSlashIfNeeded = ( url ) =>
 {
-    return url.replace(/\/$/, "");
-}
+    const urlObj = parse( url );
+    const urlPath = urlObj.path;
+    let extName;
+
+    if ( urlPath )
+    {
+        extName = path.extname( urlPath );
+    }
+
+    let slashedUrl = url;
+
+    if ( urlPath && !urlObj.hash && !extName && !urlPath.endsWith( '/' ) )
+    {
+        slashedUrl += '/';
+    }
+
+    return slashedUrl;
+};
 
 export const removeTrailingHash = ( url ) =>
-{
-    return url.replace(/\#$/, "");
-}
+    url.replace( /\#$/, '' );
 
 export const removeTrailingRedundancies = ( url ) =>
 {
-    let newUrl = url.replace( /index\.html$/, '');
-    newUrl = newUrl.replace( /index$/, '');
-    // newUrl = removeTrailingSlash( newUrl );
+    let newUrl = url.replace( /index\.html$/, '' );
+    newUrl = newUrl.replace( /index$/, '' );
+    newUrl = removeTrailingSlash( newUrl );
     newUrl = removeTrailingHash( newUrl );
 
     // loop until clean
-    if( newUrl === url )
+    if ( newUrl === url )
     {
-        return newUrl
+        return newUrl;
     }
-    else
-    {
-        return removeTrailingRedundancies( newUrl );
-    }
-}
 
-const getProtocolPosition = ( url, inputProtocol  ) =>
+    return removeTrailingRedundancies( newUrl );
+};
+
+const getProtocolPosition = ( url, inputProtocol ) =>
 {
     const fullProto = '://';
     const shortProto = ':';
@@ -56,16 +73,18 @@ const getProtocolPosition = ( url, inputProtocol  ) =>
     }
 
     return protocolPos;
-}
+};
 /**
  * Takes input and adds requisite url portions as needed, comparing to package.json defined
- * protocols, or defaulting to http
+ * protocols, or defaulting to http.
+ *
+ * Strips trailling slashes/hashses for clarity in the address bar
  * @param  {String} input address bar input
  * @return {String}       full url with protocol and any trailing (eg: http:// / .com)
  */
-export const makeValidUrl = ( input ) =>
+export const makeValidAddressBarUrl = ( input ) =>
 {
-    if( !input )
+    if ( !input )
     {
         return 'about:blank';
         logger.warn( 'url must be a string' );
@@ -77,7 +96,7 @@ export const makeValidUrl = ( input ) =>
 
     let finalProtocol;
     let everythingAfterProtocol = '';
-    let protocolPos = getProtocolPosition( input, inputProtocol );
+    const protocolPos = getProtocolPosition( input, inputProtocol );
 
     if ( validProtocols.includes( inputProtocol ) )
     {
@@ -105,7 +124,7 @@ export const makeValidUrl = ( input ) =>
 
         everythingAfterProtocol = `localhost:${parsedURL.hostname}/${everythingAfterProtocol}`;
     }
-    else if ( inputProtocol)
+    else if ( inputProtocol )
     {
         // TODO: Show error page for bad urls.
         return removeTrailingRedundancies( input );